requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
celery
redis
python-multipart
pydantic-settings
httpx

